requirements.txt
openai==1.54.4
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
